--- code/start.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-./bin/main
