+++ test/curl-commands.sh
@@ -1,8 +1,20 @@
 #Insert a new text passage post  
-curl -XPOST -H 'Content-Type: application/json' -d '{"user-id": 101, "type": "text","active": true,  "text-message" : "Honey Roasted Peanuts" }' http://127.0.0.1:3000/v1/posts 
+curl -XPOST -H 'Content-Type: application/json' -d \
+ '{"user-id": 101, "type": "text","active": true,  "text-message" : "Honey Roasted Peanuts" }' http://127.0.0.1:3000/v1/posts 
+
+curl -XPOST -H 'Content-Type: application/json' -d \
+ '{"user-id": 101, "type": "text","active": true,  "text-message" : "Honey Roasted Peanuts" }' http://127.0.0.1:8082/test 
+
+
+curl -XPOST -H 'Content-Type: application/json' -d \
+ '{"user-id": 101, "type": "text","active": true,  "text-message" : "Honey Roasted Peanuts", "created-at": "Nov 25 16:00:51 PST 2015",  "updated-at": "Nov 25 16:00:51 PST 2015" }' http://127.0.0.1:3000/v1/posts
+
  
+curl -XGET -H 'Content-Type: application/json' http://127.0.0.1:9200/postindex/?pretty=true
+
 #create a new image post, and store image into Mongodb
-curl -XPOST -H 'Content-Type: application/json' -d '{"user-id": 201, "type": "image","active": true,  "title" : "mylogo",  "comment" : "This is an image file" , "link" : "image=@/Users/huazhang/git/post-message-service/test/mylogo.jpg"}' http://127.0.0.1:3000/v1/posts 
+curl -XPOST -H 'Content-Type: application/json' -d \
+'{"user-id": 201, "type": "image","active": true,  "title" : "mylogo", "link" : "image=@/Users/huazhang/git/post-message-service/test/mylogo.jpg", name: "logo", "comment" : "This is an image file" }' http://127.0.0.1:3000/v1/posts 
 
 curl \
   -F "user-id=201" \
